% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genYY_AR.R
\name{genYY_AR}
\alias{genYY_AR}
\title{Function to simulate longitudinal network data under LLSM-AR model}
\usage{
genYY_AR(Phi, Tau, Beta, TT, dd, nn)
}
\description{
\code{genYY_AR} generate longitudinal networks under LLSM-AR model
}
\details{
\code{genYY_AR} generate longitudinal networks under LLSM-AR model
}

