% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getBeta}
\alias{getBeta}
\alias{getLS}
\alias{getLSlsm}
\alias{getLikelihood}
\alias{getPhi}
\alias{getSigma}
\title{Helper function to get the samples from the posterior chain of the parameter after burnin and thining}
\usage{
getBeta(object, burnin = 0, thin = 1)
}
\arguments{
\item{object}{object of class LLSM returned from either \code{lsm}, \code{lsmCOV}, 
\code{llsmRW}, \code{llsmRWcov} or \code{llsmAR}.}

\item{burnin}{number of samples from the posterior chain that needs to be discarded}

\item{thin}{argument specifying the thinin parameter for the chain}
}
\description{
\code{getIntercept} returns the samples from the posterior chain of the parameter after burnin and thining
}
\details{

}

