% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLikelihood.R
\name{getLikelihood}
\alias{getLikelihood}
\title{Function to get the MCMC chain of Intercept after burnin and thinning}
\usage{
getLikelihood(object, burnin = 0, thin = 1)
}
\arguments{
\item{object}{fitted model object of class 'LLSM'}

\item{burnin}{numeric value to specify the number of draws that must be discarded as burnin. Default value
is 0.}

\item{thin}{numeric value that is used to specify the step at which MCMC chain must be kept. Default value
is 1.}
}
\description{
\code{getLikelihood} returns the likelihood of the data at posterior chain of the MCMC in object of class
'LLSM'
}
\details{
\code{getLikelihood} returns the likelihood of the data at posterior chain of the MCMC in object of class
'LLSM'
}

