% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModeLS.R
\name{getModeLS}
\alias{getModeLS}
\title{Function to get the posterior Mode of the latent space positions at each time point}
\usage{
getModeLS(latent.space.pos)
}
\arguments{
\item{latent.space.pos}{matrix of dimension n by d by K, where K is the length of the chain}
}
\description{
\code{getModeLS} returns the posterior Mode of the MCMC chain the latent space positions 
at fixed time point
}
\details{
\code{getModeLS} returns the posterior Mode of the MCMC chain the latent space positions 
at fixed time point. Current version of \code{getModeLS} works for 2-dimensional latent space only
}

