% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPhi.R
\name{getPhi}
\alias{getPhi}
\title{Function to get the MCMC chain of Phi after burnin and thinning}
\usage{
getPhi(object, burnin = 0, thin = 1)
}
\arguments{
\item{object}{fitted model object of class 'LLSM'}

\item{burnin}{numeric value to specify the number of draws that must be discarded as burnin. Default value
is 0.}

\item{thin}{numeric value that is used to specify the step at which MCMC chain must be kept. Default value
is 1}
}
\description{
\code{getPhi} returns the posterior chain of the VAR parameter in LLSM-AR models from the object of class
'LLSM'
}
\details{
\code{getPhi} returns the posterior chain of the VAR parameter in LLSM-AR models from the object of class
'LLSM'
}

