% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{likelihood}
\alias{likelihood}
\title{Function to compute the logistic log-likelihood in the latent space model}
\usage{
likelihood(Y, Z, intercept)
}
\arguments{
\item{Y}{the adjacency matrix of binary ties information}

\item{Z}{the matrix of latent positions}

\item{intercept}{the intercept in the model}
}
\description{
\code{likelihood} returns the logistic log-likelihood in LSM
}
\details{
\code{likelihood} returns the logistic log-likelihood in LSM 
(for both directed and undirected networks)
}

