\name{LLSM-package}
\alias{LLSM-package}
\alias{LLSM}
\docType{package}
\title{
\packageTitle{LLSM}
}
\description{
\packageDescription{LLSM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{LLSM}
\packageIndices{LLSM}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{LLSM}

Maintainer: \packageMaintainer{LLSM}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}

