% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procrustes.R
\name{procrustes}
\alias{procrustes}
\title{Function to do Procrustes transformation on a matrix to 
to a fixed target}
\usage{
procrustes(Z00, C, z)
}
\arguments{
\item{C}{centering matrix}

\item{z}{matrix on which Procrustes transformation is done}
}
\description{
Function to do Procrustes transformation on a 
matrix to a fixed target
}
\details{
Function to do Procrustes transformation on a 
matrix to a fixed target
}

